app/requirements.txt
fastapi==0.85.1
pydantic==1.10.2
uvicorn==0.15.0
sqlalchemy==1.4.9
pymysql==1.0.2
cryptography==3.4.7
alembic==1.5.8
passlib==1.7.4
python-jose==3.2.0
python-multipart==0.0.5
python-dotenv==0.17.0
requests==2.27.1
mysql_connector_python==8.0.28
aiofiles==0.7.0
pyfcm==1.5.4
haversine==2.5.1
pytest==7.1.2
python-logstash==0.4.8
python-dateutil==2.8.2
boto3==1.24.88